initial review branch commit
After updating to Xcode 8.2.1 some code was altered to reflect changes
in Swift. This initial commit reflects those changes that xCode
automatically created. At this time, I am unsure which version of Xcode
Shubh wants me to use, so I am creating a new branch to ensure that no
work is lost. This branch will also reflect the review of each file and
method that Shubh wants me to do so that I can understand the
underlying function of the entire program.